# src/main/java/org/example/lab2/dto/AuthRequest.java
package org.example.lab2.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class AuthRequest {
    @NotBlank
    private String username;

    @NotBlank
    private String password;
}